requirements.txt
python-dotenv==1.1.0
requests==2.32.3
psycopg2-binary==2.9.10
gspread==6.2.1
google-auth==2.40.3
Faker
